Add the New Save Panel
In this commit, I've added a new save panel accessible from the Menu Bar. Users can name and mark a query as a favorite while saving it. They can choose to save the query in the default directory or a user-selected directory. If the default directory option is unchecked, a BFilePanel opens to set the directory. The save button is disabled until a name is entered in the BTextControl. Currently, there's no option to change the default directory, but this may be added later.