Incremental Find Panel - Refactor
I made sure to use more safe ways of writing the code in this commit. This includes working as much as possible with C++ objects that follow RAII versus just using raw C-style objects.

Apart from this, I also made a few small changes such as rewriting most functions to return a status_t value so that I could use these return values in the future for better checking of errors to increase robustness.

A few things are still left that need to be done:
	1) Resizing the entire Incremental Search Panel according to the columns present. This includes setting a minimum size based on the columns.
	2) Adding the ability to combine the columns using the "OR" combinator
	3) Trimming down the code in a few sections to make it shorter and cleaner

Apart from these short-term goals, some long-term goals with this panel are:
	1) Find a way to debounce queries so that the button can be discarded and searches can be executed simply by typing. I did explore this avenue but it seems to be bombarding the back-end with too many messages at the moment.
	2) Find a way to store and restore the columns accurately so that even the incremental queries can be saved :)